+++ Implementation/include/World.h
@@ -6,7 +6,7 @@
 class World
 {
 public:
-	explicit World(std::unique_ptr<sf::RenderWindow> window, const double physicsTimestep) : window_(std::move(window)),
+	explicit World(std::unique_ptr<sf::RenderWindow> window, const float physicsTimestep) : window_(std::move(window)),
 	physicsTimestep_(physicsTimestep){}
 
 	void AddEntity(std::unique_ptr<Entity>);
@@ -19,13 +19,13 @@ private:
 	std::unique_ptr<sf::RenderWindow> window_;
 
 	Vector2D cameraPos_;
-	double physicsTimestep_;
+	float physicsTimestep_;
 
 	/*int physicFramePerSeconds_;
 	int framePerSeconds_;*/
 
-	void Update(double);
-	static void UpdatePhysics(double);
+	void Update(float);
+	static void UpdatePhysics(float);
 
 	void ClickEvent(sf::Event);
 };
